requirements.txt
flask==3.0.3
opencv-python>=4.9.0
numpy>=1.26.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
twilio>=9.0.0
# YOLOv8 & Torch (heavy; comment out if using mock mode only)
ultralytics>=8.3.0
